+++ engine/entity/NXTerrain.h
@@ -28,7 +28,7 @@ namespace NX {
 	public:
 		virtual void Render(struct RenderParameter &renderer) override;
 		virtual ENTITY_TYPE GetEntityType() override;
-		virtual void OnTick(const int dwMillSeconds) override;
+		virtual void OnTick(const float dwMillSeconds) override;
 
 	public:
 		float GetMaxRangeByXAxis() const;
